+++ src/main/java/com/tristankechlo/toolleveling/config/primitives/LongValue.java
@@ -1,0 +1,16 @@
+package com.tristankechlo.toolleveling.config.primitives;
+
+import com.google.gson.JsonElement;
+
+public class LongValue extends NumberValue<Long> {
+
+    public LongValue(String name, Long defaultValue, Long minValue, Long maxValue) {
+        super(name, defaultValue, minValue, maxValue);
+    }
+
+    @Override
+    Long getAsType(JsonElement jsonElement) {
+        return jsonElement.getAsLong();
+    }
+
+}