+++ src/main/java/com/tristankechlo/toolleveling/config/util/IConfigValue.java
@@ -1,4 +1,4 @@
-package com.tristankechlo.toolleveling.config.values;
+package com.tristankechlo.toolleveling.config.util;
 
 import com.google.gson.JsonObject;