+++ src/main/java/com/tristankechlo/toolleveling/network/ClientNetworkHandler.java
@@ -2,7 +2,7 @@ package com.tristankechlo.toolleveling.network;
 
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
-import com.tristankechlo.toolleveling.ToolLeveling;
+import com.tristankechlo.toolleveling.ToolLevelingClient;
 import com.tristankechlo.toolleveling.client.screen.ItemValueScreen;
 import com.tristankechlo.toolleveling.config.util.ConfigSyncing;
 import com.tristankechlo.toolleveling.utils.Names.NetworkChannels;
@@ -33,7 +33,7 @@ public final class ClientNetworkHandler {
         buf.writeBlockPos(pos);
         Identifier identifier = Registry.ENCHANTMENT.getId(enchantment);
         if (identifier == null) {
-            ToolLeveling.LOGGER.warn("Error while encoding the packet for " + NetworkChannels.SET_ENCHANTMENT_LEVEL.toString());
+            ToolLevelingClient.LOGGER.warn("Error while encoding the packet for " + NetworkChannels.SET_ENCHANTMENT_LEVEL);
         }
         buf.writeIdentifier(identifier);
         buf.writeInt(level);
@@ -46,10 +46,10 @@ public final class ClientNetworkHandler {
         client.execute(() -> {
             boolean check = ConfigSyncing.deserializeConfig(identifier, json);
             if (!check) {
-                ToolLeveling.LOGGER.error("Config " + identifier + " could not be loaded");
+                ToolLevelingClient.LOGGER.error("Config '{}' could not be loaded", identifier);
                 throw new RuntimeException("Config " + identifier + " could not be loaded");
             } else {
-                ToolLeveling.LOGGER.info("Config " + identifier + " recieved and loaded.");
+                ToolLevelingClient.LOGGER.info("Config '{}' recieved and loaded.", identifier);
             }
         });
     }