+++ src/main/java/com/tristankechlo/toolleveling/utils/Utils.java
@@ -1,7 +1,7 @@
 package com.tristankechlo.toolleveling.utils;
 
 import com.google.common.collect.ImmutableMap;
-import com.tristankechlo.toolleveling.config.ItemValues;
+import com.tristankechlo.toolleveling.config.ItemValueConfig;
 import com.tristankechlo.toolleveling.config.ToolLevelingConfig;
 import net.minecraft.enchantment.Enchantment;
 import net.minecraft.enchantment.Enchantments;
@@ -19,17 +19,17 @@ public final class Utils {
 
     public static long getEnchantmentUpgradeCost(Enchantment enchantment, int level) {
         double globalModifier = ToolLevelingConfig.globalUpgradeCostMultiplier.getValue();
-        double specificModifier = ToolLevelingConfig.enchantmentUpgradeCostModifier.getMap().getOrDefault(enchantment, 1.0D);
-        long minCost = ToolLevelingConfig.minUpgradeCost.getValue();
+        double specificModifier = ToolLevelingConfig.enchantmentUpgradeCostModifier.getValue().getOrDefault(enchantment, 1.0D);
+        long minCost = ToolLevelingConfig.minimumUpgradeCost.getValue();
         // formula: (0.87x^2 + 300x) * specificModifier * globalModifier
         return (long) Math.max(minCost, ((0.87 * level * level) + (300 * level)) * specificModifier * globalModifier);
     }
 
     public static long getItemWorth(Item item) {
-        if (ItemValues.itemValues.containsKey(item)) {
-            return ItemValues.itemValues.get(item);
+        if (ItemValueConfig.itemValues.getValue().containsKey(item)) {
+            return ItemValueConfig.itemValues.getValue().get(item);
         }
-        return ToolLevelingConfig.defaultItemWorth.getValue();
+        return ItemValueConfig.defaultItemWorth.getValue();
     }
 
     public static long getItemWorth(ItemStack stack) {
@@ -43,16 +43,16 @@ public final class Utils {
     public static boolean isEnchantmentAtCap(Enchantment enchantment, int level) {
         short globalEnchantmentCap = ToolLevelingConfig.globalEnchantmentCap.getValue();
         if (globalEnchantmentCap > 0) {
-            if (ToolLevelingConfig.enchantmentCaps.getMap().containsKey(enchantment)) {
-                short enchantmentCap = ToolLevelingConfig.enchantmentCaps.getMap().get(enchantment);
+            if (ToolLevelingConfig.enchantmentCaps.getValue().containsKey(enchantment)) {
+                short enchantmentCap = ToolLevelingConfig.enchantmentCaps.getValue().get(enchantment);
                 if (enchantmentCap < globalEnchantmentCap) {
                     return level >= enchantmentCap;
                 }
             }
             return level >= globalEnchantmentCap;
         }
-        if (ToolLevelingConfig.enchantmentCaps.getMap().containsKey(enchantment)) {
-            return (level >= ToolLevelingConfig.enchantmentCaps.getMap().get(enchantment));
+        if (ToolLevelingConfig.enchantmentCaps.getValue().containsKey(enchantment)) {
+            return (level >= ToolLevelingConfig.enchantmentCaps.getValue().get(enchantment));
         }
         return false;
     }